requirements.txt
python-gitlab==2.5.0
PyYAML==5.3.1
requests==2.24.0
